llvm: update tag to e5d51463
 - Replace deprecated `Optional::getValue()` with `Optional::value()`
   and deprecated `Optional::hasValue()` with `Optional::has_value()`
 - Fixed warnings from comparison of signed and unsigned values by
   casting unsigned values to signed values
 - Fixed warning from returning an `APInt` from a lambda

+++ lib/Dialect/gml_st/transforms/tiling.cc
@@ -50,7 +50,7 @@ Value createTile(OpBuilder &b, Location loc, Value superset, ValueRange ivs,
   SmallVector<int64_t> staticSizes;
   SmallVector<Value> dynamicSizes;
   staticSizes.reserve(rank);
-  for (int64_t i = 0; i < rank; ++i) {
+  for (int64_t i = 0; i < static_cast<int64_t>(rank); ++i) {
     // If the dimension is perfectly tiled, use the statically known tile size.
     if (tileSizes[i] == 1 || (supersetShape[i] != ShapedType::kDynamicSize &&
                               supersetShape[i] % tileSizes[i] == 0)) {
@@ -165,8 +165,9 @@ LogicalResult tileUniqueFunctionResult(
 
   // All nested tiles must be of the same rank as the source value.
   int64_t rank = sourceTy.getRank();
-  if (llvm::any_of(nestedTileSizes,
-                   [&](auto it) { return it.size() != rank; })) {
+  if (llvm::any_of(nestedTileSizes, [&](auto it) {
+        return static_cast<int64_t>(it.size()) != rank;
+      })) {
     return failure();
   }
